% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_coh.R
\name{aggregate_coh}
\alias{aggregate_coh}
\title{aggregate_coh}
\usage{
aggregate_coh(ls)
}
\arguments{
\item{ls_coh}{list - list of coh items}
}
\value{
returns the aggregated coh values devided by n entries which are !=NA for each place.
}
\description{
aggregates the coh values for each place.
}
\examples{
# load librarys
require(LinguGeo)
}
\author{
Andreas Schönberg
}
