% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shannon.R
\name{calc_shannon}
\alias{calc_shannon}
\title{calc_shannon}
\usage{
calc_shannon(df, colname_coh = "coh")
}
\arguments{
\item{df}{data.frame - with at least the coherence values and x y coordinates}

\item{colname_coh}{character - column name with the coherence values. Default="coh" (provided by 'LinguGeo::coherence'}
}
\value{
returns the data.frame and adds a column with the Shannon entropy values
}
\description{
calculates the local measure of Shannon entropy
}
\examples{
# load data
hunde_utm <- read.csv(system.file("extdata","Hunde_classified_utm.csv",package = "LinguGeo"),row.names = 1)

head(hunde_utm)

# calculate coherence
coh <- coherence(dat = hunde_utm,cl = "new",xcord = 1,ycord = 2,nk = 5,na_value = NULL,develop = F)

# calculate shannon
shn <- calc_shannon(coh)
}
\author{
Andreas Schönberg
}
