% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherenceIndexOLD.R
\name{coh_re}
\alias{coh_re}
\title{coh_re - Function for plotting of PAM correction.}
\usage{
coh_re(x, y, long = NULL, lat = NULL, nk, reverse)
}
\arguments{
\item{df}{data.frame - with column "CTR" for control values, "PAM" for pam values
and optional "PAM_corr" for corrected PAM values.}

\item{al}{numeric - a vector of numeric values on the x axes to draw lines. Default=0}

\item{yl}{numeric - ylim adjustment. Default=c(0,2)}

\item{sortby}{character - column name to reorder the inoput datafram 'df'. IF set the argument 'al' will not be used.}

\item{abs_dif}{boolean - IF TRUE plots the differences as absolute values (no negative values). Default=TRUE.}

\item{titel}{character - desired maintitel for the plot . Default="PAMcorrection".}
}
\value{
returns a plot containing the original PAM (orange) and Control (blue) values along with the difference (grey).
If corrected PAM values are available those will be plotted in red along with the differnece between the corrected PAM and COntrol values.
Additionally prints the total Difference along with mean and standart divation for PAM or if given PAM corrected and Control
}
\description{
visualizes the correction of PAM values by correction value.
}
\note{
For visualization purposes the difference between the corrected PAM and Control is used as absolute values (negative values are plotted as positive).
Further the mean and sd is calculated for the absolute values.
}
\examples{
# load data

}
\author{
Andreas Schönberg
}
