% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherenceIndex3.R
\name{coherence}
\alias{coherence}
\title{coherence}
\usage{
coherence(
  dat,
  cl,
  xcord = NULL,
  ycord = NULL,
  nk = NULL,
  develop = T,
  na_value = NA
)
}
\arguments{
\item{dat}{data.frame - with at least one column for classification and x y coordinates for the places}

\item{cl}{character - column name containing classes}

\item{xcord}{numeric - column position with x-coordinates}

\item{ycord}{numeric - column position with y-coordinates}

\item{nk}{numeric - amount of nearest neighbors to use. Maximum is 19}

\item{develop}{boolean - If TRUE will return the full data.frame with all steps. Default=FALSE}

\item{na_value}{NA, NULL, numeric or character - the value which indicates NA in the 'class' column of the data.frame (see details)}
}
\value{
returns the coherence values for each data point
}
\description{
calculates the coherence by using a nearest neighbor approach
}
\note{
no notes
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
