% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherenceIndex.R
\name{coherenceIndex}
\alias{coherenceIndex}
\title{coherenceIndex}
\usage{
coherenceIndex(dat, cl, xcord = NULL, ycord = NULL, nk = NULL, reverse)
}
\arguments{
\item{dat}{data.frame or SpatialPointData.frame.}

\item{cl}{charater - name of column containing the classes}

\item{xcord}{numeric - column with x-coordinates}

\item{ycord}{numeric - column with y-coordinates}

\item{nk}{numeric - amount of nearest neighbors to use. Maximum is 19.}

\item{reverse}{boolean - develop. If TRUE saves to df in order, else in reverse.}
}
\value{
returns
}
\description{
Calculates the Coherence Index by using nearest Neighbors.
}
\note{
For
}
\examples{
# load librarys
require(rgdal)
require(spatstat)
# load data
utm <- readOGR(system.file("extdata","hunde_utm.shp",package = "LinguGeo"))
# take a look
head(utm)
# calculate Coherence Index for "hunde" with "type"
CohInd <- coherenceIndex(utm,utm$type,6,7,nk=5,reverse = F)
}
\author{
Andreas Schönberg
}
