% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherenceIndex2.R
\name{coherenceIndex2}
\alias{coherenceIndex2}
\title{coherenceIndex2}
\usage{
coherenceIndex2(dat, cl, xcord = NULL, ycord = NULL, nk = NULL)
}
\arguments{
\item{dat}{data.frame.}

\item{cl}{charater - name of column containing the classes}

\item{xcord}{numeric - column with x-coordinates}

\item{ycord}{numeric - column with y-coordinates}

\item{nk}{numeric - amount of nearest neighbors to use. Maximum is 19.}
}
\value{
returns the coherence index for each data point including local variation.
}
\description{
Calculates the Coherence Index by using an nearest neighbors approach.
}
\note{
no notes
}
\examples{
# load librarys
require(LinguGeo)
require(spatstat)
require(stringr)

# load data
csv <- read.csv(system.file("extdata","hunde_utm.csv",package = "LinguGeo"))
# take a look
head(csv)

# reclassify and trim all rows with no class (0)
csv <- phenmn_class2(data=csv,colname = "hunde",
                     pat_exp = c("nd|nt","ng|n.g","nn|n$")
                     ,cl_to = c("nd"   ,"ng"    ,"nn")
                     ,trim=T)

# calculate Coherence Index for "hunde" with "type"
coh2 <- coherenceIndex2(csv,csv$class,2,3,nk=5)

# compare to older version without local variation
coh <- coherenceIndex(csv,csv$class,2,3,nk=5)
}
\author{
Andreas Schönberg
}
