% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherenceIndex3.R
\name{coherenceIndex3}
\alias{coherenceIndex3}
\title{coherenceIndex3}
\usage{
coherenceIndex3(
  dat,
  cl,
  xcord = NULL,
  ycord = NULL,
  nk = NULL,
  develop = T,
  na_value = NA
)
}
\arguments{
\item{dat}{data.frame.}

\item{cl}{charater - name of column containing the classes}

\item{xcord}{numeric - column with x-coordinates}

\item{ycord}{numeric - column with y-coordinates}

\item{nk}{numeric - amount of nearest neighbors to use. Maximum is 19.}

\item{develop}{boolean - If TRUE will return the full dataframe with all steps. Default=FALSE}

\item{na_value}{NA, NULL, numeric or character - the value which indicates NA in the 'class' column of the dataframe.(see details).}
}
\value{
returns the coherence index for each data point including local variation.
}
\description{
Calculates the Coherence Index by using an nearest neighbors approach.
}
\note{
no notes
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
