% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCRS.R
\name{convertCRS}
\alias{convertCRS}
\title{convertCRS}
\usage{
convertCRS(df, pos_x, pos_y, src_proj, trg_proj, colnames_trg = NULL)
}
\arguments{
\item{df}{data.frame - with at least x and y coordinates}

\item{pos_x}{numeric - column position with X-coordinates}

\item{pos_y}{numeric - column position with Y-coordinates}

\item{src_proj}{character - CRS string for the source CRS}

\item{trg_proj}{character - CRS string for the traget CRS}

\item{colnames_trg}{character - desired column names for the converted coordinates.
IF NULL the default names will be "trg_x", "trg_y". Default=NULL.}
}
\value{
returns the data.frame with additional columns for the desired x y coordinates.
}
\description{
convert the CRS for the dataframe
}
\examples{
# load data
maurer_wgs <- openxlsx::read.xlsx(xlsxFile =system.file("extdata","Subset_Maurer_wgs.xlsx",package = "LinguGeo"))

# check column positions and structure
head(maurer_wgs) # Coordinates are in Long/Lat Format
str(maurer_wgs) # Coordinates are "character"

# define source and target projection
src <- "+proj=longlat +datum=WGS84 +no_defs" # WGS 84 (Long / Lat)
trg <- "+proj=utm +zone=32 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs" # UTM 32

# add utm coordinates to dataframe
maurer_utm <- convertCRS(df = maurer_wgs,
                         pos_x = 3,
                         pos_y = 4,
                         src_proj = src,
                         trg_proj = trg)

head(maurer_utm)

# define new colnames
maurer_utm2 <- convertCRS(df = maurer_wgs,
                          pos_x = 3,
                          pos_y = 4,
                          src_proj = src,
                          trg_proj = trg,
                          colnames_trg = c("utm_x","utm_y") )

head(maurer_utm2)
}
\author{
Andreas Schönberg
}
