% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinguGeo-Package.R
\docType{data}
\name{hunde.csv - data}
\alias{hunde.csv - data}
\title{Example csv Dataset}
\format{
\code{"utils::read.csv"}
}
\description{
A csv with long, lat and the spelling of "hunde" as well as classified spelling by phenomen.
}
