% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinguGeo-Package.R
\docType{data}
\name{hunde_utm.shp - data}
\alias{hunde_utm.shp - data}
\title{Example shp Dataset}
\format{
\code{"rgdal::readOGR"}
}
\description{
A shp projected in UTM32N with long, lat and the spelling of "hunde" as well as classified spelling by phenomen.
}
