% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinguGeo-Package.R
\docType{package}
\name{LinguGeo - package}
\alias{LinguGeo - package}
\title{LinguGeo - Package for linguistic Geography}
\description{
The \code{LinguGeo} package is used for methods of linguistic Geography.
}
\details{
LinguGeo Package
}
\author{
Andreas Schönberg
}
\keyword{package}
