% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplacesNA.R
\name{mergeMultiPlaces}
\alias{mergeMultiPlaces}
\title{mergeMultiPlaces}
\usage{
mergeMultiPlaces(df, pos_x, pos_y, col, col_ls = NULL)
}
\arguments{
\item{df}{data.frame - with at least x y coordinates and one data column}

\item{pos_x}{numeric - column position with x-coordinates}

\item{pos_y}{numeric - column position with y-coordinates}

\item{col}{numeric - column position with class information}

\item{col_ls}{list, numeric - list with numeric column position which should be merged in
addition to the "class" column. If col_ls==NULL only the class column will be merged. (see details). Default=NULL}
}
\value{
returns the data.frame with only the unique places.
}
\description{
merges information of places with identical coordinates
}
\details{
soon
}
\note{
Only useful for classified data!
}
\examples{
# soon
}
\author{
Andreas Schönberg
}
