% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMultiPlaces2.R
\name{mergeMultiPlaces2}
\alias{mergeMultiPlaces2}
\title{mergeMultiPlaces2}
\usage{
mergeMultiPlaces2(df, pos_x, pos_y)
}
\arguments{
\item{df}{data.frame.}

\item{pos_x}{numeric - column with x-coordinates}

\item{pos_y}{numeric - column with y-coordinates}
}
\value{
returns only unique places.
}
\description{
Merges data and class information for places with multiple entires.
}
\note{
no notes
}
\examples{
# load librarys
require(LinguGeo)
require(spatstat)
require(stringr)
}
\author{
Andreas Schönberg
}
