% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMultiPlaces3.R
\name{mergeMultiPlaces3}
\alias{mergeMultiPlaces3}
\title{mergeMultiPlaces3}
\usage{
mergeMultiPlaces3(df, pos_x, pos_y, col)
}
\arguments{
\item{df}{data.frame.}

\item{pos_x}{numeric - column with x-coordinates}

\item{pos_y}{numeric - column with y-coordinates}

\item{col}{numeric - column with "class"}
}
\value{
returns only unique places.
}
\description{
Merges class information for places with multiple entires.
}
\details{
uses pasted x and y coordinates to search for multiple entries and
merges all entires in class for dublicates
}
\note{
Only useful for classified data!
}
\examples{
# load librarys
require(LinguGeo)
require(spatstat)
require(stringr)
}
\author{
Andreas Schönberg
}
