% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplacesNA.R
\name{mergeMultiPlacesNA}
\alias{mergeMultiPlacesNA}
\title{mergeMultiPlacesNA}
\usage{
mergeMultiPlacesNA(df, pos_x, pos_y, col, col_ls = NULL)
}
\arguments{
\item{df}{data.frame.}

\item{pos_x}{numeric - column with x-coordinates}

\item{pos_y}{numeric - column with y-coordinates}

\item{col}{numeric - column with "class"}

\item{col_ls}{list, numeric - list with numeric column position which should be merged in
addition to the "class" column. If col_ls==NULL only the class column will be merged. (see details). Default=NULL.}
}
\value{
returns only unique places.
}
\description{
Merges information for dublicated places with.
}
\details{
soon
}
\note{
Only useful for classified data!
}
\examples{
# soon
}
\author{
Andreas Schönberg
}
