% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenmn_class2.R
\name{phenmn_class}
\alias{phenmn_class}
\title{phenmn_class}
\usage{
phenmn_class(data, colname, pat_exp, cl_to, set_NA = F)
}
\arguments{
\item{data}{data.frame - with at least x y coordinates and one data column}

\item{colname}{character - name of column containing the data to be classified}

\item{pat_exp}{character - expressions which will be used for pattern matching}

\item{cl_to}{character - desired names for the classes. Must be of same length as 'pat_exp'}

\item{set_NA}{boolean - if TRUE will convert all class '0' to NA (see notes)}
}
\value{
Returns the data.frame with an additional column containing the classes.
}
\description{
semi-automatic classification of language items by regular expression
}
\note{
The data may match with multiple patterns which could lead to more than 1 class for a single item.
If single items contain multiple classes the function will provide a warning.
All entries which does not match any of the given pattern will be assigned to class "0". If 'set_NA' == T all class 0 entries will be converted to NA.
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
