% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenmn_class2.R
\name{phenmn_class2}
\alias{phenmn_class2}
\title{phenmn_class2}
\usage{
phenmn_class2(data, colname, pat_exp, cl_to, set_NA = F)
}
\arguments{
\item{data}{data.frame - with at least x y coordinates and one data column.}

\item{colname}{character- name of column containing the data which should be classified.}

\item{pat_exp}{character - expressions which will be used for pattern matching}

\item{cl_to}{character - desired names for the classes. Must be of same length as 'pat_exp'.}

\item{set_NA}{boolean - if TRUE will convert all class '0' to NA (see notes)}
}
\value{
returns the dataframe with an additional column containing the classes.
}
\description{
Semi-automatic phenomenon classification for language items by expressions.
}
\note{
The data may match with multiple patterns which could lead to more than 1 class for a single items.
If single items contain multiple classes the function will provide a warning.
All entries which does not match any of the given pattern will be assigned to class "0". If 'set_NA' == T all class 0 entires will be converted to NA.

TEXT describing other cases! coming soon
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
