% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agt.R
\name{plot_agt}
\alias{plot_agt}
\title{plot_agt}
\usage{
plot_agt(agt)
}
\arguments{
\item{agt}{data.frame - with aggregated coh values computed by 'LinguGeo::aggregate_coh'}
}
\value{
returns a ggplot2 based visualization for the aggregated coh values..
}
\description{
Specialised ggplot2 based plot function to visualize the aggregated coh values.
}
\examples{
# load librarys
require(LinguGeo)
}
\author{
Andreas Schönberg
}
