% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coh2.R
\name{plot_coh}
\alias{plot_coh}
\title{plot_coh}
\usage{
plot_coh(df, est_dens = F, coh_th = NULL, bins = 10)
}
\arguments{
\item{df}{data.frame - in the format provided by 'LinguGeo::coherence'}

\item{est_dens}{boolean - if TRUE performs a KDE and adds the contours to the plot. Default = FALSE}

\item{coh_th}{numeric - threshold for the coherence index to use for KDE (only used if 'est_dens' = T). Default=NULL}

\item{bins}{numeric - number of bins for KDE (only used if 'est_dens' = T). Default=10}
}
\value{
A plot with the language item classes in different colors and the coherence index as alpha values.
}
\description{
ggplot2 based plot function in order to plot the coherence values
}
\examples{
# load data
hunde_utm <- read.csv(system.file("extdata","Hunde_classified_utm.csv",package = "LinguGeo"),row.names = 1)

head(hunde_utm)

# calculate coherence
coh <- coherence(dat = hunde_utm,cl = "new",xcord = 1,ycord = 2,nk = 5,na_value = NULL,develop = F)

# calculate coherence and set class "0" to NA
coh_NA <- coherence(dat = hunde_utm,cl = "new",xcord = 1,ycord = 2,nk = 5,na_value = 0,develop = F)

# plot coherence
plot_coh(coh)

# plot coherence with KDE
plot_coh(coh,est_dens = T,coh_th = 0.9,bins = 5)

# plot coherence with NA
plot_coh(coh_NA)
}
\author{
Andreas Schönberg
}
