% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shannon.R
\name{plot_shannon}
\alias{plot_shannon}
\title{plot_shannon}
\usage{
plot_shannon(df, add_data = F)
}
\arguments{
\item{df}{data.frame - in the format provided by 'LinguGeo::calc_shannon' (see details).}

\item{add_data}{Boolean - If FALSE will plot the shannon entropy, if TRUE will plot the Language Item classes and use alpha based on shannon entropy values. Default=F}
}
\value{
A plot with the shannon entropy or the language item classes in different colors and the shannon entropy as alpha.
}
\description{
Specialized plot function for the Shannon entropy.
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
