% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_coh.R
\name{aggregate_coh}
\alias{aggregate_coh}
\title{aggregate_coh}
\usage{
aggregate_coh(ls)
}
\arguments{
\item{ls}{list - list of dataframes computed by 'LinguGeo::coh'}
}
\value{
returns the aggregated coherence values divided by n entries which are !=NA for each site
}
\description{
aggregates the coherence values for each site
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
