% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shannon.R
\name{calc_shannon}
\alias{calc_shannon}
\title{calc_shannon}
\usage{
calc_shannon(df, colname_coh = "coh")
}
\arguments{
\item{df}{data.frame.}

\item{colname_coh}{character - name of column with the coherence values. Default="nrm" (provided by 'LinguGeo::CoherenceIndex'.}
}
\value{
returns the data.frame and adds a column with shannon entropy values.
}
\description{
Calculates the shannon entropy.
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
