% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherenceIndex.R
\name{coherenceIndex}
\alias{coherenceIndex}
\title{coherenceIndex}
\usage{
coherenceIndex(
  dat,
  cl,
  xcord = NULL,
  ycord = NULL,
  nk = NULL,
  reverse,
  develop = F
)
}
\arguments{
\item{dat}{data.frame.}

\item{cl}{charater - name of column containing the classes}

\item{xcord}{numeric - column with x-coordinates}

\item{ycord}{numeric - column with y-coordinates}

\item{nk}{numeric - amount of nearest neighbors to use. Maximum is 19.}

\item{reverse}{boolean - develop. If TRUE saves to df in order, else in reverse.}
}
\value{
return
}
\description{
Calculates the Coherence Index by using an nearest neighbors approach.
}
\note{
For
}
\examples{
# load librarys
require(LinguGeo)
require(spatstat)
# load data
csv <- read.csv(system.file("extdata","class_hunde_utm.csv",package = "LinguGeo"))
# take a look
head(csv)
# calculate Coherence Index for "hunde" with "type"
CohInd <- coherenceIndex(csv,csv$class,2,3,nk=5,reverse = F)
head(CohInd)
}
\author{
Andreas Schönberg
}
