% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCRS.R
\name{convertCRS}
\alias{convertCRS}
\title{convertCRS}
\usage{
convertCRS(df, pos_x, pos_y, src_proj, trg_proj, colnames_trg = NULL)
}
\arguments{
\item{df}{data.frame - with at least x and y coordinates}

\item{pos_x}{numeric - column position with X-coordinates}

\item{pos_y}{numeric - column position with Y-coordinates}

\item{src_proj}{character - CRS string for the source CRS}

\item{trg_proj}{character - CRS string for the traget CRS}

\item{colnames_trg}{character - desired column names for the converted coordinates.
IF NULL the default names will be "trg_x", "trg_y". Default=NULL.}
}
\value{
returns the data.frame with additional columns for the desired x y coordinates.
}
\description{
convert the CRS for the dataframe
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
