% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agt.R
\name{plot_agt}
\alias{plot_agt}
\title{plot_agt}
\usage{
plot_agt(df_agt)
}
\arguments{
\item{df_agt}{data.frame - with aggregated coherence values computed by 'LinguGeo::aggregate_coh'}
}
\value{
returns a ggplot2 based visualization for the aggregated coherence values
}
\description{
ggplot2 based plot function in order to visualize the aggregated coherence values
}
\examples{
# load librarys
require(LinguGeo)
}
\author{
Andreas Schönberg
}
