% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agt2.R
\name{plot_agt2}
\alias{plot_agt2}
\title{plot_agt2}
\usage{
plot_agt2(df_agt)
}
\arguments{
\item{df_agt}{data.frame - with aggregated coherence values computed by 'LinguGeo::aggregate_coh'}
}
\value{
returns a ggplot2 based visualization for the aggregated coherence values.
}
\description{
Specialized ggplot2 based plot function to visualize the aggregated coherence values.
}
\examples{
# load librarys
require(LinguGeo)
}
\author{
Andreas Schönberg
}
