% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coh.R
\name{plot_coh}
\alias{plot_coh}
\title{plot_coh}
\usage{
plot_coh(df, degree = 1, inverse = F)
}
\arguments{
\item{df}{data.frame in the format provided by 'coherence_index' (see deteails).}

\item{degree}{charater - name of column containing the strings}

\item{invers}{boolean -}
}
\value{
A plot with the language item classes in different colors and the coherence index as alpha values.
}
\description{
Specialized plot function for the Coherence index.
}
\examples{
# load data
coh <- read.csv(system.file("extdata","coh_utm.csv",package = "LinguGeo"))
# take a look
head(coh)
# plot the coherence index and laguage items
plot_coh(coh)
# adjust intensity
plot_coh(coh,degree=0.2)
# plot inverse
plot_coh(coh,inverse=T)
}
\author{
Andreas Schönberg
}
