% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coh.R
\name{plot_coh}
\alias{plot_coh}
\title{plot_coh}
\usage{
plot_coh(df, degree = 1, inverse = F, est_dens = F, coh_th = NULL, bins = 10)
}
\arguments{
\item{df}{data.frame in the format provided by 'coherence_index' (see deteails).}

\item{degree}{charater - name of column containing the strings}

\item{est_dens}{boolean - If TRUE performs a KDE and adds the contours to the plot. Default = FALSE}

\item{coh_th}{numeric - Threshold for the coherence index to use for KDE (only used if 'est_dens' = T). Default=NULL}

\item{bins}{numeric - Number of countour bins for KDE (only used if 'est_dens' = T). Default=10}

\item{invers}{boolean -}
}
\value{
A plot with the language item classes in different colors and the coherence index as alpha values.
}
\description{
Specialized plot function for the Coherence index.
}
\examples{
# load librarys
require(LinguGeo)
require(dplyr)
# load data
coh <- read.csv(system.file("extdata","coh_utm.csv",package = "LinguGeo"))
# take a look
head(coh)
# plot the coherence index and language items
plot_coh(coh)
# adjust intensity
plot_coh(coh,degree=0.2)
# plot inverse
plot_coh(coh,inverse=T)
# plot with KDE
plot_coh(coh,est_dens=T,coh_th=0.8)
}
\author{
Andreas Schönberg
}
