% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shannon.R
\name{plot_shannon}
\alias{plot_shannon}
\title{plot_shannon}
\usage{
plot_shannon(df, add_data = F)
}
\arguments{
\item{df}{data.frame - in the format provided by 'LinguGeo::calc_shannon' (see details)}

\item{add_data}{boolean - if FALSE will plot the Shannon entropy, if TRUE will plot the language item classes and use alpha based on Shannon entropy values. Default=FALSE}
}
\value{
a plot with the Shannon entropy or the language item classes in different colors and the Shannon entropy as alpha
}
\description{
ggplot2 based plot function in order to plot the local measure of Shannon entropy
}
\examples{
# load librarys
}
\author{
Andreas Schönberg
}
